Backend/requirements/requirement.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astroid==3.3.5
async-timeout==5.0.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
defusedxml==0.8.0rc2
diff-match-patch==20241021
dill==0.3.9
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1.6
django-allauth==65.2.0
django-autoslug==1.9.9
django-ckeditor==6.7.2
django-cors-headers==4.6.0
django-countries==7.6.1
django-environ==0.11.2
django-filter==24.3
django-import-export==4.3.3
django-jalali==7.2.0
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
django-taggit==6.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-yasg==1.21.8
et_xmlfile==2.0.0
Faker==37.1.0
Flask==3.0.3
idna==3.10
inflection==0.5.1
isort==5.13.2
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.2.0
Jinja2==3.1.4
kombu==5.4.2
Markdown==3.7
MarkupSafe==3.0.2
mccabe==0.7.0
networkx==3.4.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
phonenumbers==8.13.48
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytx==0.5.10
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
simplejwt==2.0.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tablib==3.7.0
tomlkit==0.13.2
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.0
whitenoise==6.8.2
